+++ media/main.js
@@ -10,7 +10,7 @@ let errorTimeout;
 function sendMessage(message) {
   vscode.postMessage(message);
 }
-
+let authorName = "the code author";
 Sentry.init({
   dsn: "https://48cab31c3ca44781a5be625ec226b48a@o1207913.ingest.sentry.io/6341224",
 
@@ -50,7 +50,7 @@ $(document).ready(function () {
       "<button class='run-watermelon'>Run Watermelon</button><br/>"
     );
     $("#ghHolder").append(
-      "<button class='help-link'>Get help on Slack</button>"
+      `<button class='help-link'>Get help from ${authorName}</button>`
     );
     $(".run-watermelon").on("click", (event) => {
       sendMessage({ command: "run" });
@@ -176,7 +176,7 @@ $(document).ready(function () {
       sendMessage({ command: "run" });
     });
     $("#ghHolder").append(
-      "<button class='help-link' >Get help on Slack</button>"
+      `<button class='help-link' >Get help from ${authorName}</button>`
     );
     $(".help-link").on("click", (event) => {
       sendMessage({ command: "open-link", link: "https://app.slack.com" });
@@ -200,6 +200,9 @@ $(document).ready(function () {
         break;
       case "error":
         setReceivedError(message.error.errorText);
+        break;
+      case "author":
+        authorName = message.author;
         break;
     }
   });