+++ CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to the "watermelon" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.4]
+- Added github as an image source to CSP
+- Added webflow to image CSP
+- Will now log the extension version to the console on startup
+- The get help button now shows the top committer name instead of "on Slack"
+- Fixed an error where the command would break even when results were fetched
 
-## [1.1.2]
+## [1.1.3]
 - Fixed some instances of the "Get help on Slack" running search
 - Added command to the context menu (right click to use)
 - Added tooltips to guide the user